refactor: GiftRanking 컴포넌트 내 상태관리 커스텀 훅으로 분리
